refactor: Phase 5 [5.4] migrate data/ — move etl/create_fold, etl/mtl_input, poi_dataset (BPR)
