+++ pipelines/embedding/dgi.pipe.py
@@ -15,7 +15,7 @@ from configs.paths import Resources, EmbeddingEngine
 from configs.model import InputsConfig
 from embeddings.dgi.dgi import create_embedding
 from embeddings.dgi.preprocess import preprocess_dgi
-from etl.mtl_input.builders import generate_category_input, generate_next_input_from_poi
+from data.inputs.builders import generate_category_input, generate_next_input_from_poi
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
