+++ src/etl/mtl_input/loaders.py
@@ -1,52 +1,14 @@
+"""Shim — canonical location is data.inputs.loaders (Phase 5).
+
+This module will be removed at the end of Phase 6.
 """
-Data loading classes with caching.
+import warnings as _warnings
 
-Moved from etl/embedding_fusion.py for better organization.
-"""
+_warnings.warn(
+    "etl.mtl_input.loaders is deprecated; use data.inputs.loaders instead. "
+    "This shim will be removed at the end of Phase 6.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-from typing import Dict
-import pandas as pd
-import gc
-
-from configs.paths import IoPaths, EmbeddingEngine
-from configs.embedding_fusion import EmbeddingSpec
-
-
-class EmbeddingLoader:
-    """
-    Loads and caches embedding DataFrames.
-
-    Caching avoids redundant I/O when multiple tasks use the same embedding source.
-    """
-
-    def __init__(self, state: str):
-        """
-        Initialize loader for a specific state.
-
-        Args:
-            state: State name (e.g., 'florida', 'alabama')
-        """
-        self.state = state
-        self._cache: Dict[EmbeddingEngine, pd.DataFrame] = {}
-
-    def load(self, spec: EmbeddingSpec) -> pd.DataFrame:
-        """
-        Load embedding DataFrame, using cache if available.
-
-        Args:
-            spec: Embedding specification
-
-        Returns:
-            DataFrame with embedding data
-        """
-        if spec.engine not in self._cache:
-            df = IoPaths.load_embedd(self.state, spec.engine)
-            self._cache[spec.engine] = df
-            print(f"  Loaded {spec.engine.value}: {df.shape}")
-
-        return self._cache[spec.engine]
-
-    def clear_cache(self):
-        """Free memory by clearing cache."""
-        self._cache.clear()
-        gc.collect()
+from data.inputs.loaders import *  # noqa: F401, F403
